Add support to firing business events (WIP)
EventLogWriter has a new endpoint called fireBusinessEvent. It behaves as the others, but sets dataOp to null (as it's not
relevant for business events).

The private implementation of createEventLog was changed to behave nicely with nulls.

A new implementation of EventPayload called BusinessEventPayload simply returns null for getDataType (also irrelevant for
business events). In the end it's a simple difference but might improve end-user experience when creating the payloads.

The rest behave as before. The payload for business events will be included in a "data" property, which ties the schema
of the business event to expect such property. That might be a deal breaker for some..

Simple test written to test the behavior of the new endpoint, together with helper fixtures.