Improve payload creation for business-driven events
The data-change-event-specific fields are not included anymore for business events.
A different payload wrapper object (NakadiBusinessEvent) was created to hold only
the information that matters: metadata and data, the latter being extracted to the
top level of the JSON object during serialization (so the user is free to just use
its own payload class freely, not needing to forcefully set a "data" property as
before.