Extract event type from payload
Signature of EventLogWriter changed to accept an event type and a payload. This way we are able to remove the type
from inside the payload completely (which makes more sense as it's not part of the data but rather the channel to
where we send the data).

EventLog layer untouched. Tests adjusted.

Missing README changes.