remove duplicated embedded database in our example app
Previously, in the tests we had each two embedded postgres instances started up,
one as the `embeddedPostgres` bean and one as part of the `dataSource` bean.

In `@Configuration` classes calling a different bean class to create your bean will
be caught by a proxy and give the existing bean, but our application main class is
not a `@Configuration` class. Instead, we can use a method parameter to inject the
bean created by the other method.

My IDE also removed a superfluous cast in the same class, not sure why it was there.